\name{-class}
\docType{class}
\alias{-class}
\alias{Spatial}
\alias{print.summary.Spatial}
\alias{summary,Spatial-method}
\alias{plot,Spatial,missing-method}

\title{Class ""}
\description{ }
\section{Objects from the Class}{
 meaningful }
\section{Slots}{
  \describe{
    \item{\code{}:}{Object of class \code{"matrix"}; }
  }
}
\section{Methods}{
  \describe{
    \item{}{\code{signature(obj = "")}:  }
    \item{}{\code{signature(obj = "")}:  }
    \item{}{\code{signature(obj = "")}:  }
    \item{}{\code{signature(obj = "")}:  }
    \item{}{\code{signature(obj = "")}:  }
  }
}

\usage{
Spatial(bbox, proj4string = CRS(as.character(NA)))
}

\arguments{
\item{}{}
\item{}{}
\item{}{}
}

\author{ J. Blair Christian, \email{blair.christian@gmail.com} }
\note{  

}

\section{Warning }{
}

\seealso{
  \code{\link{SpatialPoints-class}},
  \code{\link{SpatialGrid-class}},
}
%\examples{}

\keyword{classes}
