% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization-Functions.R
\name{unnormalizeData_spt}
\alias{unnormalizeData_spt}
\title{Unnormalize Time–Intensity Data}
\usage{
unnormalizeData_spt(dataInput)
}
\arguments{
\item{dataInput}{A list as returned by \code{\link{normalizeData_spt}},
containing \code{timeIntensityData} and \code{dataScalingParameters}.}
}
\value{
A list with component:
\item{timeIntensityData}{data.frame of original \code{time} and
\code{intensity} values.}
}
\description{
Reverses the transformation applied by \code{\link{normalizeData_spt}},
restoring raw time and intensity values from scaled data.
}
\examples{
norm <- normalizeData_spt(data.frame(time=0:10, intensity=rnorm(11)))
unnormalizeData_spt(norm)
}
