% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublesigmoidal-Fit-Functions.R
\name{doubleSigmoidalFitFormula_spt}
\alias{doubleSigmoidalFitFormula_spt}
\title{Double‐sigmoidal fit formula with baseline h0}
\usage{
doubleSigmoidalFitFormula_spt(
  x,
  finalAsymptoteIntensityRatio,
  maximum,
  slope1Param,
  midPoint1Param,
  slope2Param,
  midPointDistanceParam,
  h0
)
}
\arguments{
\item{x}{Numeric vector.}

\item{finalAsymptoteIntensityRatio}{Ratio of second plateau to first.}

\item{maximum}{Maximum plateau height.}

\item{slope1Param}{First slope (>0).}

\item{midPoint1Param}{First midpoint (>0).}

\item{slope2Param}{Second slope (>0).}

\item{midPointDistanceParam}{Distance to second midpoint (>0).}

\item{h0}{Baseline at x = 0.}
}
\value{
Numeric vector of same length as \code{x}.
}
\description{
Evaluates the full 5‐parameter “rise+fall” model at \code{x}.
}
